pyproject.toml
[tool.poetry]
name = "madr"
version = "0.1.0"
description = "FastZero - MADR"
authors = ["clcosta <clcostaf@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.3"
alembic = "^1.13.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
taskipy = "^1.13.0"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
blue = "^0.9.1"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
config = 'pre-commit install'
format = 'blue .'
pre_test = 'task format'
test = 'pytest -s -x --cov=madr -vv'
post_test = 'coverage html'
run = 'fastapi dev madr'
migrate = 'alembic revision --autogenerate'
migrate-run = 'alembic upgrade head'
migrate-revert = 'alembic downgrade -1'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
